Moved ExportDataPostProcessing() to SourceDB interface
