+++ yb_migrate/cmd/common.go
@@ -18,9 +18,7 @@ package cmd
 import (
 	"fmt"
 	"io/ioutil"
-	"os"
 	"os/exec"
-	"path/filepath"
 	"regexp"
 	"strconv"
 	"strings"
@@ -193,53 +191,6 @@ func GetTableRowCount(filePath string) map[string]int64 {
 
 	log.Infof("tableRowCountMap: %v", tableRowCountMap)
 	return tableRowCountMap
-}
-
-func ExportDataPostProcessing(source *srcdb.Source, exportDir string, tablesProgressMetadata *map[string]*utils.TableProgressMetadata) {
-	if source.DBType == "oracle" || source.DBType == "mysql" {
-		// empty - in case of oracle and mysql, the renaming is handled by tool(ora2pg)
-	} else if source.DBType == "postgresql" {
-		renameDataFiles(tablesProgressMetadata)
-	}
-
-	saveExportedRowCount(exportDir, tablesProgressMetadata)
-}
-
-func renameDataFiles(tablesProgressMetadata *map[string]*utils.TableProgressMetadata) {
-	for _, tableProgressMetadata := range *tablesProgressMetadata {
-		oldFilePath := tableProgressMetadata.InProgressFilePath
-		newFilePath := tableProgressMetadata.FinalFilePath
-		if utils.FileOrFolderExists(oldFilePath) {
-			err := os.Rename(oldFilePath, newFilePath)
-			if err != nil {
-				utils.ErrExit("renaming data file for table %q after data export: %v", tableProgressMetadata.TableName, err)
-			}
-		}
-	}
-}
-
-func saveExportedRowCount(exportDir string, tablesMetadata *map[string]*utils.TableProgressMetadata) {
-	filePath := exportDir + "/metainfo/flags/tablesrowcount"
-	file, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
-	if err != nil {
-		panic(err)
-	}
-	defer file.Close()
-	fmt.Println("exported num of rows for each table")
-	fmt.Printf("+%s+\n", strings.Repeat("-", 65))
-	fmt.Printf("| %30s | %30s |\n", "Table", "Row Count")
-	sortedKeys := utils.GetSortedKeys(tablesMetadata)
-	for _, key := range sortedKeys {
-		tableMetadata := (*tablesMetadata)[key]
-		fmt.Printf("|%s|\n", strings.Repeat("-", 65))
-
-		targetTableName := strings.TrimSuffix(filepath.Base(tableMetadata.FinalFilePath), "_data.sql")
-		actualRowCount := tableMetadata.CountLiveRows
-		line := targetTableName + "," + strconv.FormatInt(actualRowCount, 10) + "\n"
-		file.WriteString(line)
-		fmt.Printf("| %30s | %30d |\n", key, actualRowCount)
-	}
-	fmt.Printf("+%s+\n", strings.Repeat("-", 65))
 }
 
 //setup a project having subdirs for various database objects IF NOT EXISTS
